+++ odra-wasm-client/src/types.rs
@@ -1,20 +1,17 @@
-mod access_rights;
-mod cl;
-mod deploy;
-mod digest;
+mod address;
+mod bigint;
+mod bytes;
+mod public_key;
 mod signature_response;
 mod transaction;
-mod uref_addr;
+mod uref;
 mod verbosity;
 
-pub use cl::{
-    address::Address,
-    bigint::{U128, U256, U512},
-    bytes::Bytes,
-    public_key::PublicKey,
-    uref::URef
-};
-pub use deploy::Deploy;
-pub use signature_response::SignatureResponse;
+pub use address::Address;
+pub use bigint::{U128, U256, U512};
+pub use bytes::Bytes;
+pub use public_key::PublicKey;
+pub(crate) use signature_response::SignatureResponse;
 pub use transaction::{Transaction, TransactionHash};
+pub use uref::URef;
 pub use verbosity::Verbosity;
