+++ odra-wasm-client/src/types/bytes.rs
