--- odra-wasm-client/src/types/cl/mod.rs
@@ -1,5 +1,0 @@
-pub mod address;
-pub mod bigint;
-pub mod bytes;
-pub mod public_key;
-pub mod uref;
