+++ odra-wasm-client/src/types/signature_response.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
 #[derive(Debug, Deserialize, Serialize)]
-pub struct SignatureResponse {
+pub(crate) struct SignatureResponse {
     cancelled: bool,
     #[serde(rename = "signatureHex")]
     signature_hex: Option<String>,
@@ -18,16 +18,16 @@ impl SignatureResponse {
         self.signature_hex.clone().unwrap_or_default()
     }
 
-    pub fn get_signature(&self) -> Vec<u8> {
-        let signature = self.signature.clone().unwrap_or_default();
-        let mut signature_vec = vec![0; signature.len()];
-        for (key, value) in &signature {
-            if let Ok(index) = key.parse::<usize>() {
-                if index < signature_vec.len() {
-                    signature_vec[index] = *value;
-                }
-            }
-        }
-        signature_vec
-    }
+    // pub fn get_signature(&self) -> Vec<u8> {
+    //     let signature = self.signature.clone().unwrap_or_default();
+    //     let mut signature_vec = vec![0; signature.len()];
+    //     for (key, value) in &signature {
+    //         if let Ok(index) = key.parse::<usize>() {
+    //             if index < signature_vec.len() {
+    //                 signature_vec[index] = *value;
+    //             }
+    //         }
+    //     }
+    //     signature_vec
+    // }
 }
